# [performance] WIP Parallelization of route edge and wait costing iteration