--- a/peartree/paths.py
+++ b/peartree/paths.py
@@ -57,7 +57,8 @@ def load_feed_as_graph(feed: ptg.gtfs.feed,
                        connection_threshold: float=50.0,
                        walk_speed_kmph: float=4.5,
                        interpolate_times: bool=True,
-                       impute_walk_transfers: bool=False):
+                       impute_walk_transfers: bool=False,
+                       use_multiprocessing: bool=True):
     """
     Convert a feed object into a NetworkX Graph, connect to an existing
     NetworkX graph if one is supplied
@@ -96,6 +97,9 @@ def load_feed_as_graph(feed: ptg.gtfs.feed,
     impute_walk_transfers : bool
         A flag to indicate whether to add in walk connections between nodes
         that are close enough, as measured using connection_trheshold
+    use_multiprocessing: bool
+        A flag to indicate whether or not to leverage multiprocessing where
+        available to attempt to speed up trivially parallelizable operations.
 
     Returns
     -------
@@ -120,7 +124,8 @@ def load_feed_as_graph(feed: ptg.gtfs.feed,
      wait_times_by_stop) = generate_summary_graph_elements(feed,
                                                            start_time,
                                                            end_time,
-                                                           interpolate_times)
+                                                           interpolate_times,
+                                                           use_multiprocessing)
 
     # This is a flag used to check if we need to run any additional steps
     # after the feed is returned to ensure that new nodes and edge can connect