--- a/peartree/summarizer.py
+++ b/peartree/summarizer.py
@@ -1,5 +1,6 @@
 from typing import Dict, List, Tuple, Union
 
+import dask.bag as dbag
 import numpy as np
 import pandas as pd
 import partridge as ptg
@@ -322,13 +323,54 @@ def generate_edge_and_wait_values(
     else:
         stop_times = feed.stop_times.copy()
 
+    route_analyzer = RouteProcessor(
+        target_time_start,
+        target_time_end,
+        ftrips,
+        stop_times,
+        feed.stops.copy())
+
+    route_id_list = dbag.from_sequence(feed.routes.route_id.head())
+    res = route_id_list.map(route_analyzer.generate_route_costs).compute()
+
     all_edge_costs = None
     all_wait_times = None
-    for i, route in feed.routes.iterrows():
-        log('Processing on route {}.'.format(route.route_id))
+    for tst_sub, edge_costs in res:
+        # Add to the running total for wait times in this feed subset
+        if all_wait_times is None:
+            all_wait_times = tst_sub
+        else:
+            all_wait_times = all_wait_times.append(tst_sub)
 
+        # Add to the running total in this feed subset
+        if all_edge_costs is None:
+            all_edge_costs = edge_costs
+        else:
+            all_edge_costs = all_edge_costs.append(edge_costs)
+
+    return (all_edge_costs, all_wait_times)
+
+
+class RouteProcessor(object):
+
+    def __init__(
+            self,
+            target_time_start: int,
+            target_time_end: int,
+            feed_trips: pd.DataFrame,
+            stop_times: pd.DataFrame,
+            all_stops: pd.DataFrame):
+
+        # Initialize common parameters
+        self.target_time_start = target_time_start
+        self.target_time_end = target_time_end
+        self.trips = feed_trips.copy()
+        self.stop_times = stop_times.copy()
+        self.all_stops = all_stops.copy()
+
+    def generate_route_costs(self, route_id: str):
         # Get all the subset of trips that are related to this route
-        trips = ftrips.loc[route.route_id]
+        trips = self.trips.loc[route_id]
 
         # Pandas will try and make returned result a Series if there
         # is only one result - prevent this from happening
@@ -336,18 +378,19 @@ def generate_edge_and_wait_values(
             trips = trips.to_frame().T
 
         # Get just the stop times related to this trip
-        st_trip_id_mask = stop_times.trip_id.isin(trips.trip_id)
-        stimes_init = stop_times[st_trip_id_mask]
+        st_trip_id_mask = self.stop_times.trip_id.isin(trips.trip_id)
+        stimes_init = self.stop_times[st_trip_id_mask]
 
         # Then subset further by just the time period that we care about
-        start_time_mask = (stimes_init.arrival_time >= target_time_start)
-        end_time_mask = (stimes_init.arrival_time <= target_time_end)
+        start_time_mask = (stimes_init.arrival_time >= self.target_time_start)
+        end_time_mask = (stimes_init.arrival_time <= self.target_time_end)
         stimes = stimes_init[start_time_mask & end_time_mask]
 
         # Report on progress if requested
         a = len(stimes_init.trip_id.unique())
         b = len(stimes.trip_id.unique())
-        log('\tReduced trips in consideration from {} to {}.'.format(a, b))
+        log('\tReduced selected trips on route {} from {} to {}.'.format(
+            route_id, a, b))
 
         trips_and_stop_times = pd.merge(trips,
                                         stimes,
@@ -355,7 +398,7 @@ def generate_edge_and_wait_values(
                                         on='trip_id')
 
         trips_and_stop_times = pd.merge(trips_and_stop_times,
-                                        feed.stops,
+                                        self.all_stops,
                                         how='inner',
                                         on='stop_id')
 
@@ -372,19 +415,7 @@ def generate_edge_and_wait_values(
                                         'wait_dir_0',
                                         'wait_dir_1']]
 
-        # Add to the running total for wait times in this feed subset
-        if all_wait_times is None:
-            all_wait_times = tst_sub
-        else:
-            all_wait_times = all_wait_times.append(tst_sub)
-
         # Get all edge costs for this route and add to the running total
         edge_costs = generate_all_observed_edge_costs(trips_and_stop_times)
 
-        # Add to the running total in this feed subset
-        if all_edge_costs is None:
-            all_edge_costs = edge_costs
-        else:
-            all_edge_costs = all_edge_costs.append(edge_costs)
-
-    return (all_edge_costs, all_wait_times)
+        return (tst_sub, edge_costs)