--- a/requirements.txt
+++ b/requirements.txt
@@ -1,5 +1,3 @@
-dask==0.17.2
-distributed==1.21.4
 fiona==1.6.1
 networkx>=2.0
 osmnx==0.6