--- a/tests/test_graph.py
+++ b/tests/test_graph.py
@@ -22,12 +22,14 @@ def test_generate_summary_graph_elements():
     start = 7 * 60 * 60
     end = 10 * 60 * 60
     interpolate_times = True
+    use_multiprocessing = True
 
     (summary_edge_costs,
      wait_times_by_stop) = generate_summary_graph_elements(feed_1,
                                                            start,
                                                            end,
-                                                           interpolate_times)
+                                                           interpolate_times,
+                                                           use_multiprocessing)
 
     # Ensure that the summary edge cost dataframe looks as it should
     ec_cols = ['edge_cost', 'from_stop_id', 'to_stop_id']