--- a/tests/test_graph.py
+++ b/tests/test_graph.py
@@ -24,52 +24,55 @@ def test_generate_summary_graph_elements():
     interpolate_times = True
     use_multiprocessing = True
 
-    (summary_edge_costs,
-     wait_times_by_stop) = generate_summary_graph_elements(feed_1,
-                                                           start,
-                                                           end,
-                                                           interpolate_times,
-                                                           use_multiprocessing)
+    # Make sure everything works the same with both multiprocessing on/off
+    for use_multiprocessing in [True, False]:
+        (summary_edge_costs,
+         wait_times_by_stop) = generate_summary_graph_elements(
+            feed_1,
+            start,
+            end,
+            interpolate_times,
+            use_multiprocessing)
 
-    # Ensure that the summary edge cost dataframe looks as it should
-    ec_cols = ['edge_cost', 'from_stop_id', 'to_stop_id']
-    for c in ec_cols:
-        assert c in summary_edge_costs.columns
+        # Ensure that the summary edge cost dataframe looks as it should
+        ec_cols = ['edge_cost', 'from_stop_id', 'to_stop_id']
+        for c in ec_cols:
+            assert c in summary_edge_costs.columns
 
-    # Make sure that all edges are unique - there are no duplicated
-    # in the returned edge dataframe (each should be its own summary)
-    f = summary_edge_costs.from_stop_id
-    t = summary_edge_costs.to_stop_id
-    z = list(zip(f, t))
-    assert len(list(set(z))) == len(z)
+        # Make sure that all edges are unique - there are no duplicated
+        # in the returned edge dataframe (each should be its own summary)
+        f = summary_edge_costs.from_stop_id
+        t = summary_edge_costs.to_stop_id
+        z = list(zip(f, t))
+        assert len(list(set(z))) == len(z)
 
-    # Ensure that the wait times dataframe looks as it should
-    wt_cols = ['avg_cost', 'stop_id']
-    for c in wt_cols:
-        assert c in wait_times_by_stop.columns
+        # Ensure that the wait times dataframe looks as it should
+        wt_cols = ['avg_cost', 'stop_id']
+        for c in wt_cols:
+            assert c in wait_times_by_stop.columns
 
-    # Sanity check edge costs
-    mask = (wait_times_by_stop.avg_cost < 0)
-    assert len(wait_times_by_stop[mask]) == 0
+        # Sanity check edge costs
+        mask = (wait_times_by_stop.avg_cost < 0)
+        assert len(wait_times_by_stop[mask]) == 0
 
-    # Make sure that there are stop ids unique
-    u = wait_times_by_stop.stop_id.unique()
-    assert len(u) == len(wait_times_by_stop)
+        # Make sure that there are stop ids unique
+        u = wait_times_by_stop.stop_id.unique()
+        assert len(u) == len(wait_times_by_stop)
 
-    # Another sanity check, we should be sure that the resulting
-    # edges list captures all the stops that were assigned null
-    # values in the fixture dataset were assigned a linearly imputed
-    # arrival and departure time and thus preserved as a stop
-    # in the edge list
+        # Another sanity check, we should be sure that the resulting
+        # edges list captures all the stops that were assigned null
+        # values in the fixture dataset were assigned a linearly imputed
+        # arrival and departure time and thus preserved as a stop
+        # in the edge list
 
-    # First get the null times mask
-    null_times = feed_1.stop_times.departure_time.isnull()
-    # And identify all unique stops from the original feed
-    null_stop_ids = feed_1.stop_times[null_times].stop_id.unique()
+        # First get the null times mask
+        null_times = feed_1.stop_times.departure_time.isnull()
+        # And identify all unique stops from the original feed
+        null_stop_ids = feed_1.stop_times[null_times].stop_id.unique()
 
-    # Now let's take the list of these null stop ids and extract
-    # all the ones from that list in the summary edge dataframe
-    mask = summary_edge_costs.from_stop_id.isin(null_stop_ids)
-    # And now we can get the stop ids out from this list
-    preserved_from_nulls = summary_edge_costs.from_stop_id[mask].unique()
-    assert len(preserved_from_nulls) == 205
+        # Now let's take the list of these null stop ids and extract
+        # all the ones from that list in the summary edge dataframe
+        mask = summary_edge_costs.from_stop_id.isin(null_stop_ids)
+        # And now we can get the stop ids out from this list
+        preserved_from_nulls = summary_edge_costs.from_stop_id[mask].unique()
+        assert len(preserved_from_nulls) == 205