This replaced work on https://github.com/kuanb/peartree/pull/51/files

From the previous PR:
Partially (incrementally) addressing issue #12

Parallelizes target route processing operation process_route_edges_and_wait_times via dask distributed which allows for modular parallelization architecture which in the future could leverage external resources (useful for large graphs, tethering together whole regions, etc.).

------

Updates unique to this new PR:

Using multiprocessing, _not_ Dask, for now. Change is incremental.

------
OLD:

Going with a Dask Bag for now. Keeping it simple for now, and can improve later.

Results from a quick test with AC Transit and the new system:

With `interpolate_times` set to `False`:
Without Dask (original method):
```
CPU times: user 40.4 s, sys: 920 ms, total: 41.3 s
Wall time: 41 s
```

With Dask:
```
CPU times: user 13.2 s, sys: 530 ms, total: 13.7 s
Wall time: 14.6 s
```

For reference, this is doing just the first 5 routes:
Without Dask:
```
CPU times: user 6.15 s, sys: 140 ms, total: 6.29 s
Wall time: 6.34 s
```

With Dask:
```
CPU times: user 13 s, sys: 440 ms, total: 13.5 s
Wall time: 14.3 s
```

Even more dramatic is when you set the time interpolation to `True`:
Without Dask: 3min 55s
With Dask: 1min 17s

From this, the initial cost of about 14.3 seconds can be seen to initialize the various Dask configurations to enable the parallelization. The upside is the significantly reduced marginal cost of each additional unique route.

Of course, a lot of this matters on the machine you are running. Allowing for access to Dask Distributed's Client will be next to do, which will enable utilizing external resources.